+++ packages/core/src/awsService/ec2/sshKeyPair.ts
@@ -4,7 +4,7 @@
  */
 import { fs } from '../../shared'
 import { ToolkitError } from '../../shared/errors'
-import { tryRun } from '../../shared/utilities/processUtils'
+import { tryRun } from '../../shared/utilities/programUtils'
 import { Timeout } from '../../shared/utilities/timeoutUtils'
 import { findAsync } from '../../shared/utilities/collectionUtils'
 import { RunParameterContext } from '../../shared/utilities/processUtils'
