+++ packages/core/src/shared/utilities/processUtils.ts
@@ -7,7 +7,6 @@ import * as proc from 'child_process'
 import * as crossSpawn from 'cross-spawn'
 import * as logger from '../logger'
 import { Timeout, CancellationError, waitUntil } from './timeoutUtils'
-import { getLogger } from '..'
 
 export interface RunParameterContext {
     /** Reports an error parsed from the stdin/stdout streams. */
@@ -368,31 +367,4 @@ export class ChildProcess {
         const pid = this.pid() > 0 ? `PID ${this.pid()}:` : '(not started)'
         return `${pid} [${this.#command} ${noparams ? '...' : this.#args.join(' ')}]`
     }
-}
-
-/**
- * Tries to execute a program at path `p` with the given args and
- * optionally checks the output for `expected`.
- *
- * @param p path to a program to execute
- * @param args program args
- * @param doLog log failures
- * @param expected output must contain this string
- */
-export async function tryRun(
-    p: string,
-    args: string[],
-    logging: 'yes' | 'no' | 'noresult' = 'yes',
-    expected?: string,
-    opt?: ChildProcessOptions
-): Promise<boolean> {
-    const proc = new ChildProcess(p, args, { logging: 'no' })
-    const r = await proc.run(opt)
-    const ok = r.exitCode === 0 && (expected === undefined || r.stdout.includes(expected))
-    if (logging === 'noresult') {
-        getLogger().info('tryRun: %s: %s', ok ? 'ok' : 'failed', proc)
-    } else if (logging !== 'no') {
-        getLogger().info('tryRun: %s: %s %O', ok ? 'ok' : 'failed', proc, proc.result())
-    }
-    return ok
 }
